Add a support listeners with union type of event
