+++ tests/common/Dispatcher/CompositeDispatcherTest.php
