+++ tests/common/Event/ClassInterface.php
