+++ tests/common/Event/ClassItself.php
