+++ tests/common/Event/StoppableEvent.php
