+++ tests/common/Listener/NonStatic.php
