+++ tests/common/Listener/WithStaticMethod.php
