+++ tests/common/Provider/CompositeTest.php
